requirements.txt
pytest==7.4.3
selene==2.0.0rc4
webdriver-manager==4.0.1
allure-pytest==2.13.2